+++ internal/controller/gitsync_controller.go
@@ -32,8 +32,8 @@ import (
 	"github.com/numaproj-labs/numaplane/internal/controller/config"
 	"github.com/numaproj-labs/numaplane/internal/git"
 	"github.com/numaproj-labs/numaplane/internal/kubernetes"
+	"github.com/numaproj-labs/numaplane/internal/shared/gitconfig"
 	"github.com/numaproj-labs/numaplane/internal/shared/logging"
-	"github.com/numaproj-labs/numaplane/internal/shared/validations"
 )
 
 // GitSyncReconciler reconciles a GitSync object
@@ -277,7 +277,7 @@ func (r *GitSyncReconciler) validate(gitSync *apiv1.GitSync) error {
 	destination := gitSync.Spec.Destination
 
 	// Validate the repositoryPath
-	if ok := validations.CheckGitURL(repositoryPath.RepoUrl); !ok {
+	if ok := gitconfig.CheckGitURL(repositoryPath.RepoUrl); !ok {
 		return fmt.Errorf("invalid remote repository url %s", repositoryPath.RepoUrl)
 	}
 	if len(repositoryPath.Name) == 0 {
@@ -291,7 +291,7 @@ func (r *GitSyncReconciler) validate(gitSync *apiv1.GitSync) error {
 	if len(destination.Cluster) == 0 {
 		return fmt.Errorf("cluster name cannot be empty")
 	}
-	if !validations.IsValidKubernetesNamespace(destination.Namespace) {
+	if !gitconfig.IsValidKubernetesNamespace(destination.Namespace) {
 		return fmt.Errorf("namespace is not a valid string for cluster %s", destination.Cluster)
 	}
