+++ internal/kubernetes/kubernetes.go
@@ -1,12 +1,14 @@
 package kubernetes
 
 import (
+	"context"
 	"fmt"
 	"path/filepath"
 
 	jsongo "github.com/json-iterator/go"
 	"go.uber.org/zap"
 	"gopkg.in/yaml.v2"
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -85,4 +87,17 @@ func ToUnstructured(manifest map[string]interface{}) (*unstructured.Unstructured
 	return &unstructured.Unstructured{
 		Object: m,
 	}, nil
+}
+
+// GetSecret  gets secret using the kubernetes client
+func (c *client) GetSecret(ctx context.Context, namespace, secretName string) (*corev1.Secret, error) {
+	secret := &corev1.Secret{}
+	key := k8sClient.ObjectKey{
+		Namespace: namespace,
+		Name:      secretName,
+	}
+	if err := c.Get(ctx, key, secret); err != nil {
+		return nil, err
+	}
+	return secret, nil
 }
