+++ internal/shared/gitconfig/gitconfig.go
@@ -1,12 +1,19 @@
-package validations
+package gitconfig
 
 import (
+	"context"
 	"fmt"
 	"log"
 	"net/url"
 	"regexp"
 
+	"github.com/go-git/go-git/v5/plumbing/transport"
+	"github.com/go-git/go-git/v5/plumbing/transport/http"
+	"github.com/go-git/go-git/v5/plumbing/transport/ssh"
 	"k8s.io/apimachinery/pkg/util/validation"
+
+	controllerconfig "github.com/numaproj-labs/numaplane/internal/controller/config"
+	"github.com/numaproj-labs/numaplane/internal/kubernetes"
 )
 
 // Valid git transports url
@@ -78,4 +85,47 @@ func IsValidKubernetesNamespace(name string) bool {
 		return true
 	}
 	return false
+}
+
+func GetAuthMethod(ctx context.Context, repoUrl string, kubeClient kubernetes.Client, namespace string, repoCred *controllerconfig.GitCredential) (transport.AuthMethod, error) {
+	scheme, err := GetTransportScheme(repoUrl)
+	if err != nil {
+		return nil, err
+	}
+	var authMethod transport.AuthMethod
+	switch scheme {
+	case "ssh", "git+ssh":
+		key := repoCred.SSHCredential.SSHKey.Key
+
+		secret, err := kubeClient.GetSecret(ctx, namespace, key)
+		if err != nil {
+			return nil, err
+		}
+		authMethod, err = ssh.NewPublicKeys("git", secret.Data["sshKey"], "")
+		if err != nil {
+			return nil, err
+		}
+
+	case "http", "https":
+		if repoCred.HTTPCredential != nil {
+			cred := repoCred.HTTPCredential
+			username := cred.Username
+			secret, err := kubeClient.GetSecret(ctx, namespace, cred.Password.Key)
+			if err != nil {
+				return nil, err
+			}
+			password := string(secret.Data["password"])
+			authMethod = &http.BasicAuth{
+				Username: username,
+				Password: password,
+			}
+		}
+	case "git", "file":
+		authMethod = nil
+
+	default:
+		return nil, fmt.Errorf("unsupported or not required authentication")
+	}
+
+	return authMethod, nil
 }
