+++ internal/shared/gitconfig/gitconfig_test.go
@@ -1,4 +1,4 @@
-package validations
+package gitconfig
 
 import (
 	"testing"
